+++ src/components/Books.js
@@ -1,16 +1,12 @@
 import React, { useState } from 'react';
 import Book from './Book';
+import NewBookForm from './NewBookForm';
 
 const Books = () => {
   const [books, setBooks] = useState([]);
 
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    const title = event.target.title.value;
-    const author = event.target.author.value;
-    const newBook = { title, author };
+  const handleAddBook = (newBook) => {
     setBooks([...books, newBook]);
-    event.target.reset();
   };
 
   return (
@@ -21,11 +17,7 @@ const Books = () => {
           <Book key={book.id} title={book.title} author={book.author} />
         ))}
       </ul>
-      <form onSubmit={handleSubmit}>
-        <input type="text" name="title" placeholder="Title" required />
-        <input type="text" name="author" placeholder="Author" required />
-        <button type="submit">Add Book</button>
-      </form>
+      <NewBookForm handleAddBook={handleAddBook} />
     </div>
   );
 };
